Is787/invitation update (#834)
Improvements on #787 

- invitation **gets deleted upon ONE use or expiration (5 days)**
- moved portal demo scripts from ``services/web/server/tests/sandbox`` ==> [scripts/demo](scripts/demo)
- create demo markdown and csv file with invitation codes

- bit of cleanup refactoring in login subsytem
- updated tests
- doc and logs
- Aiohttp server Background task loop deprecated

+++ services/web/server/tests/helpers/utils_login.py
@@ -104,4 +104,5 @@ class NewInvitation(NewUser):
         return self.confirmation
 
     async def __aexit__(self, *args):
-        await self.db.delete_confirmation(self.confirmation)
+        if await self.db.get_confirmation(self.confirmation):
+            await self.db.delete_confirmation(self.confirmation)
